+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added `All` and `Any` for `Span<T>` and `ReadOnlySpan<T>`, mimicking the corresponding methods in `System.Linq`
 - Added time-related extension methods (`Ticks`, `Milliseconds`, `Seconds`, `Minutes`, `Hours`, `Days`, and `Weeks`) to all built-in numeric types
 - Added `TimeSpan.Ago()` and `TimeSpan.FromNow()`
+- Added `Factorial` extension method for built-in numeric types
 - Added `FileInfo.GetHash<T>()`
 - Added `FileInfo.TryWriteHash<T>(Span<byte>, out int)`
 - Added `IEnumerable<TSource>.Product()` and `IEnumerable<TSource>.Product<TResult>(Func<TSource>, TResult)` for all built-in numeric types, computing the product of all (transformed) elements